Fix position_ids format and data type for FlashInfer RoPE API
Critical fixes identified:
1. **Remove .flatten() from position_ids**: FlashInfer expects 2D [batch, seq_len]
2. **Use int32 for position_ids**: Ensure correct integer type for indices
3. **Match successful model patterns**: Align with Llama/Qwen implementations

The key insight is that FlashInfer's apply_rope_with_cos_sin_cache_inplace expects:
- position_ids: [batch_size, seq_len] in int32
- query/key: [-1, dim] (flattened)
- cos_sin_cache: [max_length, rope_dim*2] in float32

+++ model_hub/deepseek.py
@@ -600,8 +600,8 @@ class DeepSeekModel(LLM):
         rms_norm_eps = getattr(self.config, 'rms_norm_eps', 1e-6)
         self.norm = RMSNorm(self.hidden_size, eps=rms_norm_eps).to(device_0)
         
-        # 初始化位置编码
-        self.position_ids = torch.arange(0, self.max_length).to(device_0)
+        # 初始化位置编码 - 确保使用正确的数据类型
+        self.position_ids = torch.arange(0, self.max_length, dtype=torch.int32).to(device_0)
         self._init_rope()
         
         # 初始化各层
